processor/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
greenlet==2.0.2
grpcio==1.56.0
grpcio-tools==1.56.0
idna==3.4
requests==2.31.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
